+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsApi.kt
@@ -98,6 +98,9 @@ internal class AnalyticsApi(
             .putOpt(FPTI_KEY_CONTEXT_TYPE, if (event.isVaultRequest) "BA-TOKEN" else "EC-TOKEN")
             .putOpt(FPTI_KEY_PAYPAL_ATTEMPTED_APP_SWITCH, event.didSdkAttemptAppSwitch)
             .putOpt(FPTI_KEY_FUNDING_SOURCE, event.fundingSource)
+            .putOpt(FPTI_KEY_IS_PURCHASE, event.isPurchase)
+            .putOpt(FPTI_KEY_IS_BILLING_AGREEMENT, event.isBillingAgreement)
+            .putOpt(FPTI_KEY_BILLING_PLAN_TYPE, event.recurringBillingPlanType)
     }
 
     @Throws(JSONException::class)
@@ -157,7 +160,9 @@ internal class AnalyticsApi(
         private const val FPTI_KEY_ERROR_DESC = "error_desc"
         private const val FPTI_KEY_CONTEXT_TYPE = "context_type"
         private const val FPTI_KEY_FUNDING_SOURCE = "funding_source"
-
+        private const val FPTI_KEY_IS_PURCHASE = "is_purchase"
+        private const val FPTI_KEY_IS_BILLING_AGREEMENT = "is_billing_agreement"
+        private const val FPTI_KEY_BILLING_PLAN_TYPE = "billing_plan_type"
         private const val FPTI_BATCH_KEY_VENMO_INSTALLED = "venmo_installed"
         private const val FPTI_BATCH_KEY_PAYPAL_INSTALLED = "paypal_installed"
         private const val FPTI_BATCH_KEY_APP_ID = "app_id"