+++ BraintreeCore/src/test/java/com/braintreepayments/api/core/AnalyticsApiUnitTest.kt
@@ -53,6 +53,9 @@ class AnalyticsApiUnitTest {
         contextId = "context-id",
         linkType = "link-type",
         isVaultRequest = true,
+        isPurchase = true,
+        isBillingAgreement = true,
+        recurringBillingPlanType = "billing_plan_type",
         startTime = 123,
         endTime = 234,
         endpoint = "endpoint",
@@ -75,6 +78,9 @@ class AnalyticsApiUnitTest {
         contextId = "context-id",
         linkType = "link-type",
         isVaultRequest = false,
+        isPurchase = true,
+        isBillingAgreement = true,
+        recurringBillingPlanType = "billing_plan_type",
         startTime = 123,
         endTime = 234,
         endpoint = "endpoint",
@@ -225,7 +231,10 @@ class AnalyticsApiUnitTest {
                                     "event_name": "${tokenizationKeyEvent.name}",
                                     "button_position": "${tokenizationKeyEvent.buttonOrder}",
                                     "context_type": "EC-TOKEN",
-                                    "paypal_app_switch_url_received": ${tokenizationKeyEvent.didPayPalServerAttemptAppSwitch}
+                                    "paypal_app_switch_url_received": ${tokenizationKeyEvent.didPayPalServerAttemptAppSwitch},
+                                    "is_purchase": ${tokenizationKeyEvent.isPurchase},
+                                    "is_billing_agreement": ${tokenizationKeyEvent.isBillingAgreement},
+                                    "billing_plan_type": "${tokenizationKeyEvent.recurringBillingPlanType}"
                                 }
                             ]
                         }
@@ -284,7 +293,10 @@ class AnalyticsApiUnitTest {
                                     "event_name": "${clientTokenEvent.name}",
                                     "button_position": "${clientTokenEvent.buttonOrder}",
                                     "context_type": "BA-TOKEN",
-                                    "paypal_app_switch_url_received": ${clientTokenEvent.didPayPalServerAttemptAppSwitch}
+                                    "paypal_app_switch_url_received": ${clientTokenEvent.didPayPalServerAttemptAppSwitch},
+                                    "is_purchase": ${tokenizationKeyEvent.isPurchase},
+                                    "is_billing_agreement": ${tokenizationKeyEvent.isBillingAgreement},
+                                    "billing_plan_type": "${tokenizationKeyEvent.recurringBillingPlanType}"
                                 }
                             ]
                         }