+++ BraintreeCore/src/test/java/com/braintreepayments/api/core/AnalyticsParamRepositoryUnitTest.kt
@@ -1,5 +1,6 @@
 package com.braintreepayments.api.core
 
+import com.braintreepayments.api.paypal.PayPalRecurringBillingPlanType
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.Before
@@ -15,6 +16,7 @@ class AnalyticsParamRepositoryUnitTest {
     private val uuid = "test-uuid"
     private val newUuid = "new-uuid"
     private val expectedFundingSource = "funding-source"
+    private val expectedBillingPlanType = PayPalRecurringBillingPlanType.RECURRING.name
 
     @Before
     fun setUp() {
@@ -26,6 +28,9 @@ class AnalyticsParamRepositoryUnitTest {
         sut.didSdkAttemptAppSwitch = true
         sut.didEnablePayPalAppSwitch = true
         sut.fundingSource = expectedFundingSource
+        sut.isPurchase = true
+        sut.isBillingAgreement = true
+        sut.recurringBillingPlanType = expectedBillingPlanType
     }
 
     @Test
@@ -46,6 +51,9 @@ class AnalyticsParamRepositoryUnitTest {
         assertEquals(true, sut.didEnablePayPalAppSwitch)
         assertEquals(true, sut.didSdkAttemptAppSwitch)
         assertEquals(expectedFundingSource, sut.fundingSource)
+        assertEquals(true, sut.isPurchase)
+        assertEquals(true, sut.isBillingAgreement)
+        assertEquals(expectedBillingPlanType, sut.recurringBillingPlanType)
 
         sut.reset()
 
@@ -54,5 +62,8 @@ class AnalyticsParamRepositoryUnitTest {
         assertNull(sut.didEnablePayPalAppSwitch)
         assertNull(sut.didSdkAttemptAppSwitch)
         assertNull(sut.fundingSource)
+        assertNull(sut.isPurchase)
+        assertNull(sut.isBillingAgreement)
+        assertNull(sut.recurringBillingPlanType)
     }
 }